requirements.txt
pandas>=2.0.0
pandas-ta>=0.3.14b
TA-Lib>=0.4.24
numpy>=1.24.0
openai>=1.0.0
groq>=0.4.0
requests>=2.31.0
ollama>=0.1.0
websockets>=11.0.0
ccxt>=4.0.0
yfinance>=0.2.0
PyYAML>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.13.0
loguru>=0.7.0